BackendServer/requirements.txt
Flask==2.3.2
flask-cors==6.0.1
mysql-connector-python==9.5.0
bcrypt==5.0.0